will fix later, process offer should run async
